BUG FIXED: Card Images showing up now + Card Links
Fixed bug causing images in cards to not show up - added size smaller than or equal to card (300px) and removed / from in front of images in src. Added card links that link to website when user clicks anywhere on the card. **Make sure to add links for TinDog and Personicle Dashboard later**